feat: Home 컴포넌트 작성, Switch 컴포넌트 적용
