+++ src/App.js
@@ -1,7 +1,8 @@
 import { hot } from 'react-hot-loader/root';
 import React from 'react';
-import { BrowserRouter as Router, Route } from 'react-router-dom';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Nav from './components/Nav';
+import Home from './components/Home';
 import TodoListTemplate from './components/TodoListTemplate';
 import { TodoProvider } from './components/TodoContext';
 
@@ -10,7 +11,10 @@ const App = () => {
     <TodoProvider>
       <Router>
         <Nav />
-        <Route path="/" component={TodoListTemplate} />
+        <Switch>
+          <Route path="/home" component={Home} />
+          <Route path="/todo" component={TodoListTemplate} />
+        </Switch>
       </Router>
     </TodoProvider>
   );
