+++ src/utils/index.js
@@ -2,3 +2,4 @@ export { default as sleep } from './sleep';
 export { default as checkStatus } from './checkStatus';
 export { default as isEmptyArr } from './isEmptyArr';
 export { default as isBlank } from './isBlank';
+export { default as countTodoStatus } from './countTodoStatus';
